# Optimize arm64 decoder