Here's a new version of the arm64 assembly decoder, with the following changes:

* Use the extended register set to cut dependencies, save on loads and make the code more readable.
* Assume unaligned loads and stores are safe and cheap.
* Use LDP/STP to copy 16 literal bytes at a time.
* Use compare-and-branch/conditional compare instructions for better code density.

Benchmark results on Raspberry Pi 4B (Go 1.17.1):

```
name                old speed      new speed       delta
UncompressPg1661-4  93.6MB/s ± 1%  101.0MB/s ± 1%   +7.88%  (p=0.000 n=10+10)
UncompressDigits-4   359MB/s ± 0%    453MB/s ± 0%  +26.08%  (p=0.000 n=10+10)
UncompressTwain-4    102MB/s ± 0%    110MB/s ± 0%   +7.74%  (p=0.000 n=10+10)
UncompressRand-4    1.13GB/s ± 1%   1.14GB/s ± 1%   +0.80%  (p=0.029 n=10+10)
```

This is a version of the code that I'd developed prior to @lizthegrey's PR #143, but never quite finished, patched up to make the diff a bit smaller (though it's still large). Sorry for not participating in that discussion.