@lizthegrey 2021-11-01
> Can you also backport this to v3? Thanks!

I'll get the backport PR up shortly