@lizthegrey 2021-11-01
https://github.com/pierrec/lz4/pull/148